+++ SND/EmulsionTarget/TTPoint.cxx
@@ -12,15 +12,16 @@ using std::endl;
 TTPoint::TTPoint() : FairMCPoint() {}
 // -------------------------------------------------------------------------
 
-TTPoint::TTPoint(Int_t trackID, Int_t detID, TVector3 pos, TVector3 mom,
-                 Double_t tof, Double_t length, Double_t eLoss, Int_t pdgcode)
+TTPoint::TTPoint(Int_t trackID, Int_t detID, const TVector3& pos,
+                 const TVector3& mom, Double_t tof, Double_t length,
+                 Double_t eLoss, Int_t pdgcode)
     : FairMCPoint(trackID, detID, pos, mom, tof, length, eLoss),
       fPdgCode(pdgcode) {}
 
 // -------------------------------------------------------------------------
 
 // -----   Destructor   ----------------------------------------------------
-TTPoint::~TTPoint() {}
+TTPoint::~TTPoint() = default;
 // -------------------------------------------------------------------------
 
 // -----   Public method Print   -------------------------------------------