+++ macro/makeCascade.py
@@ -338,6 +338,7 @@ for iev in range(args.nevgen):
         # generate a signal based on probabilities in hists i*10+8?
         ptot = ROOT.TMath.Sqrt(stack[nstack][1] ** 2 + stack[nstack][2] ** 2 + stack[nstack][3] ** 2)
         prbsig = 0.0
+        idpn = 0
         for i in range(1, id + 1):
             if stack[nstack][0] == idhist[i]:  # get hist id for this beam particle
                 idpn = 0