+++ shipgen/GenieGenerator.cxx
@@ -28,7 +28,7 @@ using std::endl;
 // important to read back number of events to give to FairRoot
 
 // -----   Default constructor   -------------------------------------------
-GenieGenerator::GenieGenerator() {}
+GenieGenerator::GenieGenerator() = default;
 // -------------------------------------------------------------------------
 // -----   Default constructor   -------------------------------------------
 Bool_t GenieGenerator::Init(const char* fileName) { return Init(fileName, 0); }
@@ -41,37 +41,68 @@ Bool_t GenieGenerator::Init(const char* fileName, const int startEvent) {
     return kFALSE;
   }
   fNuOnly = false;
-  fInputFile = TFile::Open(fileName);
   LOG(info) << "Opening input file " << fileName;
-  if (!fInputFile) {
-    LOG(fatal) << "Error opening input file.";
+  if (startEvent < 0) {
+    LOG(error) << "GenieGenerator: startEvent must be >= 0, got " << startEvent;
+    return kFALSE;
+  }
+  fInputFile = TFile::Open(fileName, "READ");
+  if (!fInputFile || fInputFile->IsZombie()) {
+    LOG(error) << "GenieGenerator: error opening input file " << fileName;
+    delete fInputFile;
+    fInputFile = nullptr;
     return kFALSE;
   }
   fTree = dynamic_cast<TTree*>(fInputFile->Get("gst"));
+  if (!fTree) {
+    LOG(error) << "GenieGenerator: cannot find tree gst in file " << fileName;
+    fInputFile->Close();
+    delete fInputFile;
+    fInputFile = nullptr;
+    return kFALSE;
+  }
   fNevents = fTree->GetEntries();
+  if (startEvent >= fNevents) {
+    LOG(error) << "GenieGenerator: startEvent " << startEvent
+               << " is out of range for " << fNevents << " entries";
+    fInputFile->Close();
+    delete fInputFile;
+    fInputFile = nullptr;
+    fTree = nullptr;
+    return kFALSE;
+  }
   fn = startEvent;
-  fTree->SetBranchAddress("Ev", &Ev);  // incoming neutrino energy
-  fTree->SetBranchAddress("pxv", &pxv);
-  fTree->SetBranchAddress("pyv", &pyv);
-  fTree->SetBranchAddress("pzv", &pzv);
-  fTree->SetBranchAddress("neu", &neu);    // incoming neutrino PDG code
-  fTree->SetBranchAddress("cc", &cc);      // Is it a CC event?
-  fTree->SetBranchAddress("nuel", &nuel);  // Is it a NUEEL event?
-  fTree->SetBranchAddress("vtxx", &vtxx);  // vertex  in SI units
-  fTree->SetBranchAddress("vtxy", &vtxy);
-  fTree->SetBranchAddress("vtxz", &vtxz);
-  fTree->SetBranchAddress("vtxt", &vtxt);
-  fTree->SetBranchAddress("El", &El);  // outgoing lepton momentum
-  fTree->SetBranchAddress("pxl", &pxl);
-  fTree->SetBranchAddress("pyl", &pyl);
-  fTree->SetBranchAddress("pzl", &pzl);
-  fTree->SetBranchAddress("Ef", &Ef);  // outgoing hadronic momenta
-  fTree->SetBranchAddress("pxf", &pxf);
-  fTree->SetBranchAddress("pyf", &pyf);
-  fTree->SetBranchAddress("pzf", &pzf);
-  fTree->SetBranchAddress("nf", &nf);      // nr of outgoing hadrons
-  fTree->SetBranchAddress("pdgf", &pdgf);  // pdg code of hadron
-
+  bool ok = true;
+  ok &= (fTree->SetBranchAddress("Ev", &Ev) >= 0);
+  ok &= (fTree->SetBranchAddress("pxv", &pxv) >= 0);
+  ok &= (fTree->SetBranchAddress("pyv", &pyv) >= 0);
+  ok &= (fTree->SetBranchAddress("pzv", &pzv) >= 0);
+  ok &= (fTree->SetBranchAddress("neu", &neu) >= 0);
+  ok &= (fTree->SetBranchAddress("cc", &cc) >= 0);
+  ok &= (fTree->SetBranchAddress("nuel", &nuel) >= 0);
+  ok &= (fTree->SetBranchAddress("vtxx", &vtxx) >= 0);
+  ok &= (fTree->SetBranchAddress("vtxy", &vtxy) >= 0);
+  ok &= (fTree->SetBranchAddress("vtxz", &vtxz) >= 0);
+  ok &= (fTree->SetBranchAddress("vtxt", &vtxt) >= 0);
+  ok &= (fTree->SetBranchAddress("El", &El) >= 0);
+  ok &= (fTree->SetBranchAddress("pxl", &pxl) >= 0);
+  ok &= (fTree->SetBranchAddress("pyl", &pyl) >= 0);
+  ok &= (fTree->SetBranchAddress("pzl", &pzl) >= 0);
+  ok &= (fTree->SetBranchAddress("Ef", &Ef) >= 0);
+  ok &= (fTree->SetBranchAddress("pxf", &pxf) >= 0);
+  ok &= (fTree->SetBranchAddress("pyf", &pyf) >= 0);
+  ok &= (fTree->SetBranchAddress("pzf", &pzf) >= 0);
+  ok &= (fTree->SetBranchAddress("nf", &nf) >= 0);
+  ok &= (fTree->SetBranchAddress("pdgf", &pdgf) >= 0);
+  if (!ok) {
+    LOG(error)
+        << "GenieGenerator: failed to bind one or more required branches";
+    fInputFile->Close();
+    delete fInputFile;
+    fInputFile = nullptr;
+    fTree = nullptr;
+    return kFALSE;
+  }
   fFirst = kTRUE;
   return kTRUE;
 }
@@ -99,12 +130,14 @@ std::vector<double> GenieGenerator::Rotate(Double_t x, Double_t y, Double_t z,
 
 // -----   Destructor   ----------------------------------------------------
 GenieGenerator::~GenieGenerator() {
-  fInputFile->Close();
-  fInputFile->Delete();
-  delete fInputFile;
+  if (fInputFile) {
+    fInputFile->Close();
+    fInputFile->Delete();
+    delete fInputFile;
+  }
 }
 // -------------------------------------------------------------------------
-void GenieGenerator::AddBox(TVector3 dVec, TVector3 box) {
+void GenieGenerator::AddBox(const TVector3& dVec, const TVector3& box) {
   dVecs.push_back(dVec);
   m_boxes.push_back(box);
   cout << "Debug GenieGenerator: " << dVec.X() << " " << box.Z() << endl;