+++ CHANGELOG.md
@@ -20,6 +20,8 @@ it in future.
 
 ### Changed
 
+* Change A scaling in makeDecay
+
 ### Fixed
 
 ### Removed